refactor DictionarySocket
