+++ app/src/main/java/org/example/dictionary/DictionaryService.java
@@ -1,0 +1,23 @@
+package org.example.dictionary;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+public class DictionaryService {
+
+    public boolean doesWordExist(String word) {
+        try {
+            URL url = new URL("https://api.dictionaryapi.dev/api/v2/entries/en/" + word);
+            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+            conn.setRequestMethod("GET");
+            conn.setConnectTimeout(5000);
+            conn.setReadTimeout(5000);
+
+            return conn.getResponseCode() == 200;
+        } catch (IOException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+}
