+++ app/src/main/java/org/example/dictionary/DictionarySocket.java
@@ -1,0 +1,25 @@
+package org.example.dictionary;
+
+import java.io.IOException;
+import java.net.ServerSocket;
+import java.net.Socket;
+
+public class DictionarySocket {
+
+    public static void initialize() throws IOException {
+        int port = 8082;
+        ServerSocket serverSocket = new ServerSocket(port);
+        System.out.println("Server started on port " + port);
+
+        new Thread(() -> {
+            while (true) {
+                try {
+                    Socket clientSocket = serverSocket.accept();
+                    new Thread(new ClientHandler(clientSocket)).start();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }).start();
+    }
+}
